refactor(use case): governance use case server and client function modules separation
